+++ client/src/component/Article.js
@@ -2,7 +2,6 @@ import React from 'react'
 import { useSelector } from 'react-redux'
 import CaseOfArticleTrue from './article/CaseOfArticleTrue'
 import CaseOfArticleFalse from './article/CaseOfArticleFalse'
-import Test from './article/Test'
 import './article.css'
 
 function Article() {
@@ -10,7 +9,7 @@ function Article() {
   const { isLogin } = state
   return (
     <div>
-      {isLogin ? <Test />: <CaseOfArticleFalse />}
+      {isLogin ? <CaseOfArticleTrue />: <CaseOfArticleFalse />}
     </div>
   )
 }
