+++ client/src/component/article/CaseOfArticleTrue.js
@@ -1,110 +1,144 @@
-import { React, useState } from 'react'
+import {React, useState } from 'react'
 import styled from 'styled-components'
-import { changeImage, detailedPostMode } from '../../actions/index'
 import { useSelector, useDispatch } from 'react-redux'
-import MapType from '../MapType'
+import { changeImage, detailedPostMode } from '../../actions'
+import MapType from '../MapType' 
 import dummydata from '../../dummy/dummydata'
+import yellowMood from '../../static/yellowMood.png'
+import greenMood from '../../static/greenMood.png'
+import redMood from '../../static/redMood.png'
+import blueMood from '../../static/blueMood.png'
+import violetMood from '../../static/violetMood.png'
 
-const Posts = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  width: 100%;
-  height: 100%;
-  overflow: scroll;
-  &:hover {
-    cursor: pointer;
-  }
-`
-// isLogin = true
-const Picture = styled.img`
-  width: 100%;
-  height: 100%;
-  :hover {
-    width: 102%;
-    height: 102%;
-  }
-`
-const PictureWrapper = styled.div`
-  @media only screen and (max-width: 1109px) {
-    width: 70%;
-    min-width: 17rem;
-  }
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 17rem;
-  height: 22rem;
-  margin: 1.5rem;
-  flex-wrap: nowrap;
-  overflow: hidden;
+
+const HoverPosts = styled.div`
+  display: table-cell;
+  vertical-align: middle;
+  text-align: center;
+  font-family: serif;
+  font-weight: bold;
+  width: 20rem;
+  height: 27rem;
+  border: 1px solid #C4C4C4;
+  border-radius: 20px;
+  margin: 1.3rem;
+  float: left;
 `
 
+const CreatedAt = styled.div`
+  text-align: left;
+  margin: 0.5rem 1rem;
+`
+
+const DetailedMood = styled.div`
+  text-align: right;
+  margin: 0.5rem;
+`
 const Mood = styled.img`
   width: 25px;
   height: 25px;
   margin-right: 6px;
 `
 
-export default function CaseOfArticleTrue() {
-  const [isHover, setIsHover] = useState(false)
-  const hideStyle = { display: 'none' }
+const PictureWrapper = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 17rem;
+  height: 20rem;
+  margin: 1.5rem;
+  flex-wrap: nowrap;
+  overflow: hidden;
+`
+const Picture = styled.img`
+  width: 17rem;
+  height: 20rem;
+  &:hover {
+    cursor: pointer;
+  }
+`
+
+export default function Test () {
+  const [isHovers, setIsHovers] = useState({
+  1: false,
+  2: false,
+  3: false,
+  4: false,
+  5: false,
+  6: false,
+  7: false,
+  8: false,
+  9: false,
+  10: false,
+  })
+  console.log(isHovers[1])
   const state = useSelector(state => state.loginReducer)
   const postState = useSelector(state => state.postReducer)
-  const pictureState = useSelector(state => state.changeImageReducer)
   const { isLogin } = state
   const { isDiary } = postState
-  const { picture } = pictureState
 
-  const moods = () => {
+
+  const moods = picture => {
+
     let mood = []
 
-    for (let i = 0; i < picture.mood.length; i++) {
-      if (picture.mood[i] === 1) {
-        mood.push(<Mood src={yellowMood} />)
+    for (let i=0;i<picture.length;i++) {
+      if (picture[i] === 1) {  
+        mood.push(<Mood src={yellowMood} />)  
       }
-      if (picture.mood[i] === 2) {
+      if (picture[i] === 2) {
         mood.push(<Mood src={greenMood} />)
       }
-      if (picture.mood[i] === 3) {
-        mood.push(<Mood src={redMood} />)
+      if (picture[i] === 3) {
+        mood.push(<Mood src={redMood} />)  
       }
-      if (picture.mood[i] === 4) {
+      if (picture[i] === 4) {
         mood.push(<Mood src={blueMood} />)
       }
-      if (picture.mood[i] === 5) {
+      if (picture[i] === 5) {
         mood.push(<Mood src={violetMood} />)
       }
     }
     return mood
-  }
+  } 
 
   const dispatch = useDispatch()
-  
-  const caseOfArticleTrue = () => {
+
+  const test = () => {
     if (isLogin) {
       if (isDiary) {
         return (
-          <Posts>
+          <>
             {dummydata.map(post => (
-              <PictureWrapper
-                key={post.id}
-                onClick={() => {
-                  dispatch(changeImage(post))
-                  dispatch(detailedPostMode())
-                }}
-              >
-                <Picture src={post.src} />
-              </PictureWrapper>
+              <>
+                <HoverPosts key={post.id}>
+                  {isHovers[post.id] ? <><CreatedAt>{post.updatedAt[0]}.{post.updatedAt[1]}.{post.updatedAt[2]}</CreatedAt>
+                  <DetailedMood>{moods(post.mood)}</DetailedMood> </> : null}
+                  <PictureWrapper  onClick={() => {
+                    dispatch(changeImage(post))
+                    dispatch(detailedPostMode())}} 
+                    onMouseEnter={() => {setIsHovers({...isHovers, [post.id]: true})}}
+                    onMouseLeave={() => {setIsHovers({...isHovers, [post.id]: false})}}>
+                    <Picture src={post.src}  />
+                  </PictureWrapper>
+                </HoverPosts>
+              </>
             ))}
-          </Posts>
+          </>         
         )
       } else {
-        return <MapType />
+        return (
+          <MapType />
+        )
       }
     }
-  }
 
+  }  
+  
+  return (
+    <>
+      {test()}
+    </>
 
-  return <>{caseOfArticleTrue()}</>
+  )
 }
