+++ server/.gitignore
@@ -49,3 +49,5 @@ ormconfig.js.txt
 
 /dummy
 /dummy/uploads
+
+/src/uploads
